requirements.txt
-e git+https://github.com/i-kiwamu/python3-oauth2.git#egg=oauth2
-e git+https://github.com/elleryq/plurk-oauth.git#egg=plurk_oauth
feedparser==5.2.1
bleach==1.4.2
python-dateutil==2.4.2
